+++ src/main/kotlin/cloud/glitchdev/rfu/constants/MessageTypes.kt
@@ -1,0 +1,11 @@
+package cloud.glitchdev.rfu.constants
+
+enum class MessageTypes(val displayName : String) {
+    HYPE("Hyperion"),
+    AUTOPET("Autopet"),
+    CATCH("Catch");
+
+    override fun toString(): String {
+        return displayName
+    }
+}